Change double to decimal in internal ValueContainer float representation
Why: FloatFunction.cs had a cast to C# float, which resulted in loss of information for a user. This resulted in the discussion of the VC Float type to be changed to `decimal` to ensure the most floating point number information is kept throughout the execution of an expression

+++ ExpressionEngine/Functions/Implementations/ConversionFunctions/IntFunction.cs
@@ -47,9 +47,9 @@ namespace ExpressionEngine.Functions.Implementations.ConversionFunctions
                     return new ValueTask<ValueContainer>(new ValueContainer(parameters[0].GetValue<int>()));
 
                 case ValueType.Float:
-                    var floatToIntVal = (int)System.Math.Round(parameters[0].GetValue<double>());
+                    var floatToIntVal = (int)System.Math.Round(parameters[0].GetValue<decimal>());
                     CheckIntMaxOrMin(floatToIntVal);
-                    return new ValueTask<ValueContainer>(new ValueContainer((int)System.Math.Round(parameters[0].GetValue<double>())));
+                    return new ValueTask<ValueContainer>(new ValueContainer((int)System.Math.Round(parameters[0].GetValue<decimal>())));
 
                 case ValueType.Boolean:
                     return new ValueTask<ValueContainer>(new ValueContainer(parameters[0].GetValue<bool>() ? 1 : 0));
